+++ cognite/client/_cognite_client.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-import warnings
-from typing import TYPE_CHECKING, Any, Dict, Optional
+from typing import Any, Dict, Optional
 
 from requests import Response
 
@@ -12,7 +11,7 @@ from cognite.client._api.data_sets import DataSetsAPI
 from cognite.client._api.diagrams import DiagramsAPI
 from cognite.client._api.entity_matching import EntityMatchingAPI
 from cognite.client._api.events import EventsAPI
-from cognite.client._api.extractionpipelines import ExtractionPipelineRunsAPI, ExtractionPipelinesAPI
+from cognite.client._api.extractionpipelines import ExtractionPipelinesAPI
 from cognite.client._api.files import FilesAPI
 from cognite.client._api.functions import FunctionsAPI
 from cognite.client._api.geospatial import GeospatialAPI
@@ -28,9 +27,6 @@ from cognite.client._api.transformations import TransformationsAPI
 from cognite.client._api.vision import VisionAPI
 from cognite.client._api_client import APIClient
 from cognite.client.config import ClientConfig, global_config
-
-if TYPE_CHECKING:
-    from cognite.client._api.datapoints import DatapointsAPI
 
 
 class CogniteClient:
@@ -96,19 +92,6 @@ class CogniteClient:
         """Perform a DELETE request to an arbitrary path in the API."""
         return self._api_client._delete(url, params=params, headers=headers)
 
-    @property  # TODO (v6.0.0): Delete this whole property
-    def datapoints(self) -> DatapointsAPI:
-        if int(self.version.split(".")[0]) >= 6:
-            raise AttributeError(  # ...in case we forget to delete this property in v6...
-                "'CogniteClient' object has no attribute 'datapoints'. Use 'time_series.data' instead."
-            )
-        warnings.warn(
-            "Accessing the DatapointsAPI through `client.datapoints` is deprecated and will be removed "
-            "in major version 6.0.0. Use `client.time_series.data` instead.",
-            DeprecationWarning,
-        )
-        return self.time_series.data
-
     @property
     def version(self) -> str:
         """Returns the current SDK version.
@@ -126,16 +109,3 @@ class CogniteClient:
             ClientConfig: The configuration object.
         """
         return self._config
-
-    @property  # TODO (v6.0.0): Delete this whole property
-    def extraction_pipeline_runs(self) -> ExtractionPipelineRunsAPI:
-        if int(self.version.split(".")[0]) >= 6:
-            raise AttributeError(
-                "'CogniteClient' object has no attribute 'extraction_pipeline_runs'. Use 'extraction_pipelines.runs' instead."
-            )
-        warnings.warn(
-            "Accessing the ExtractionPipelineRunsAPI through `client.extraction_pipeline_runs` is deprecated and will be removed "
-            "in major version 6.0.0. Use `client.extraction_pipelines.runs` instead.",
-            DeprecationWarning,
-        )
-        return self.extraction_pipelines.runs