+++ cognite/client/_api/extractionpipelines.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Dict, Optional, Sequence, Union, cast, overload
+from typing import TYPE_CHECKING, Any, Dict, Optional, Sequence, Union, overload
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
@@ -17,7 +17,6 @@ from cognite.client.data_classes import (
     TimestampRange,
 )
 from cognite.client.data_classes.extractionpipelines import StringFilter
-from cognite.client.utils._auxiliary import handle_deprecated_camel_case_argument
 from cognite.client.utils._identifier import IdentifierSequence
 
 if TYPE_CHECKING:
@@ -123,19 +122,15 @@ class ExtractionPipelinesAPI(APIClient):
         return self._list(list_cls=ExtractionPipelineList, resource_cls=ExtractionPipeline, method="GET", limit=limit)
 
     @overload
-    def create(self, extraction_pipeline: ExtractionPipeline, **kwargs: Dict[str, Any]) -> ExtractionPipeline:
+    def create(self, extraction_pipeline: ExtractionPipeline) -> ExtractionPipeline:
         ...
 
     @overload
-    def create(
-        self, extraction_pipeline: Sequence[ExtractionPipeline], **kwargs: Dict[str, Any]
-    ) -> ExtractionPipelineList:
+    def create(self, extraction_pipeline: Sequence[ExtractionPipeline]) -> ExtractionPipelineList:
         ...
 
     def create(
-        self,
-        extraction_pipeline: Union[ExtractionPipeline, Sequence[ExtractionPipeline]] = None,
-        **kwargs: Dict[str, Any],
+        self, extraction_pipeline: Union[ExtractionPipeline, Sequence[ExtractionPipeline]]
     ) -> Union[ExtractionPipeline, ExtractionPipelineList]:
         """`Create one or more extraction pipelines. <https://docs.cognite.com/api/v1/#operation/createExtPipes>`_
 
@@ -157,11 +152,6 @@ class ExtractionPipelinesAPI(APIClient):
                 >>> extpipes = [ExtractionPipeline(name="extPipe1",...), ExtractionPipeline(name="extPipe2",...)]
                 >>> res = c.extraction_pipelines.create(extpipes)
         """
-        # TODO: Remove support for old argument name in major version 6
-        extraction_pipeline = cast(
-            Union[ExtractionPipeline, Sequence[ExtractionPipeline]],
-            handle_deprecated_camel_case_argument(extraction_pipeline, "extractionPipeline", "create", kwargs),
-        )
         utils._auxiliary.assert_type(extraction_pipeline, "extraction_pipeline", [ExtractionPipeline, Sequence])
         return self._create_multiple(
             list_cls=ExtractionPipelineList, resource_cls=ExtractionPipeline, items=extraction_pipeline